+++ src/locales/en/translation.json
@@ -76,7 +76,6 @@
     "send": "Send",
     "policy": "By submitting your personal data, you agree to its processing.",
     "contactInfo": "Contacts"
-   
   },
   "Modal": {
     "title": "Quick Order",
@@ -114,7 +113,7 @@
     "lengthMax30" : "Maximum of 30 characters",
     "lengthMax100" : "Maximum of 100 characters",
     "incorrectPhone": "Please enter a valid Montenegro phone number",
-    "weakPassword": "At least one letter and one number",
+    "weakPassword": "At least one uppercase letter, one lowercase letter and one number",
     "lengthMin8" : "Minimum of 8 characters"
   },
   "Notification": {
